+++ app/renderer/index.html
@@ -1,18 +1,23 @@
-<!DOCTYPE html>
-<html>
 
+<!doctype html>
+<html lang="en" class="no-js">
 <head>
-  <title>CatWAN-Monitor</title>
-  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/latest/css/bootstrap.min.css">
-  <link rel="stylesheet" href="./style.css">
-</head>
+	<meta charset="UTF-8">
+	<meta name="viewport" content="width=device-width, initial-scale=1">
 
-<body>
-  <nav class="navbar navbar-dark bg-dark mb-3">
-    <a class="navbar-brand" href="#">
-      <img src="./logo.png" alt="" />
-    </a>
-  </nav>
+	<link href='https://fonts.googleapis.com/css?family=Open+Sans:300,400,700' rel='stylesheet' type='text/css'>
+    <!--link href="/assets/css/style.css" type="text/css" rel="stylesheet"-->
+	<link rel="stylesheet" href="/assets/css/style_index.css"> <!-- Resource style -->
+	<script src="/assets/js/modernizr.js"></script> <!-- Modernizr -->
+
+	<link href="/assets/css/style.css" type="text/css" rel="stylesheet">
+	<link rel="stylesheet" href="/assets/libs/morris.css">
+	<meta name="viewport" content="width=device-width,user-scalable=no">
+	<link rel="stylesheet" href="https://unpkg.com/leaflet@1.4.0/dist/leaflet.css"
+	integrity="sha512-puBpdR0798OZvTTbP4A8Ix/l+A4dHDD0DGqYW6RQ+9jxkRFclaxxQb/SJAWZfWAkuyeQUytO7+7N4QKrDh+drA=="
+	crossorigin=""/>
+	<title>sateliteUCI</title>
+</head>
   <div id="app"></div>
   <script src="./app.js" data-container="#app"></script>
 
@@ -25,6 +30,20 @@
       current.parentNode.insertBefore(script, current);
     }
   </script>
+  <script src="https://unpkg.com/leaflet@1.4.0/dist/leaflet.js"
+	integrity="sha512-QVftwZFqvtRNi0ZyCtsznlKSWOStnDORoefr1enyq5mVL4tmKB3S/EnC3rRJcxCPavG10IcrVGSmPh6Qw5lwrg=="
+	crossorigin=""></script>
+
+	<script src="https://ajax.googleapis.com/ajax/libs/jquery/1.9.0/jquery.min.js"></script>     
+	<script src="https://cdnjs.cloudflare.com/ajax/libs/raphael/2.1.0/raphael-min.js"></script> 
+
+	<script src="/assets/libs/morris.min.js"></script>
+  <script src="/socket.io/socket.io.js" charset="utf-8"></script>
+  <script src="/index.js" charset="utf-8"></script>
+
+<script src="/assets/js/jquery-2.1.4.js"></script>
+<script src="/assets/js/jquery.menu-aim.js"></script>
+<script src="/assets/js/main.js"></script> 
 </body>
 
 </html>
